multi-agent/requirements.txt
bedrock-agentcore>=1.0.3
strands-agents[otel]>=1.16.0
strands-agents-tools>=0.2.16
aws-opentelemetry-distro
requests>=2.31.0
boto3>=1.40.0
python-dotenv>=1.0.0
colorlog>=6.8.0
pydantic>=2.0.0
colorama>=0.4.6
rich>=13.0.0
prompt-toolkit>=3.0.0
html2text>=2024.2.26
